修改cmd命令，嵌入式使用pulseaudio控制音量

+++ SmartMusicNetCoreAudio/NetWork/MyMsg/MsgAudio.cs
@@ -10,5 +10,5 @@
     /// <summary>
     /// 声音大小0-1
     /// </summary>
-    public float audio;
+    public int audio;
 }
